+++ backend/src/App.ts
@@ -1,6 +1,5 @@
 import "reflect-metadata";
 import express from "express";
-import GraphQLMiddleware from "./graphql";
 import cors from "cors";
 import { ConnectionOptions, createConnection, getConnection } from "typeorm";
 const DB_CONN_OPTIONS: Record<
@@ -10,6 +9,10 @@ const DB_CONN_OPTIONS: Record<
 import * as socketio from "socket.io";
 import socketModule from "./socket";
 import { Server } from "http";
+import { buildSchema } from "type-graphql";
+import * as Resolver from "./resolvers";
+import { authChecker } from "./middlewares/AuthChecker";
+import { graphqlHTTP } from "express-graphql";
 
 (async () => {
   let env = "";
@@ -20,7 +23,7 @@ import { Server } from "http";
   const app = express();
   await createConnection(DB_CONN_OPTIONS[env]);
 
-  const gqMiddleware = await GraphQLMiddleware.get();
+  const gqMiddleware = await graphQLMiddleware();
 
   app.use(cors());
 
@@ -46,3 +49,20 @@ import { Server } from "http";
   });
   socketModule(io);
 })();
+
+const graphQLMiddleware = async () => {
+  const schema = await buildSchema({
+    resolvers: [
+      Resolver.UserResolver,
+      Resolver.QuestionResolver,
+      Resolver.TagResolver,
+      Resolver.AnswerResolver,
+    ],
+    authChecker: authChecker,
+  });
+
+  return graphqlHTTP({
+    schema: schema,
+    graphiql: process.env.NODE_ENV !== "production",
+  });
+};
