+++ backend/src/resolvers/index.ts
@@ -1,0 +1,6 @@
+import AnswerResolver from "./AnswerResolver";
+import QuestionResolver from "./QuestionResolver";
+import TagResolver from "./TagResolver";
+import UserResolver from "./UserResolver";
+
+export { AnswerResolver, QuestionResolver, TagResolver, UserResolver };
