+++ backend/src/services/UserService.ts
@@ -1,4 +1,4 @@
-import { User } from "../../entities/User";
+import { User } from "../entities/User";
 
 export default class UserService {
   public static async findOneUserById(id: number): Promise<User> {
